Все работает, кроме деления

+++ dlfcn/c_file/calc.h
@@ -3,9 +3,9 @@
 
 #include <stdio.h>
 
-int sum(int first, int second);
-int sub(int first, int second);
-int mul(int first, int second);
+double sum(int first, int second);
+double sub(int first, int second);
+double mul(int first, int second);
 double divis(int first, int second);
 
 #endif
