+++ dlfcn/c_file/sub.c
@@ -1,6 +1,6 @@
 #include "calc.h"
 
-int sub(int first, int second)
+double sub(int first, int second)
 {
     return (first - second);
 }
